dataspark/requirements.txt
altair==5.2.0
annotated-types==0.6.0
anyio==3.7.1
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.7.0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
docopt==0.6.2
entrypoints==0.4
exceptiongroup==1.2.0
Faker==21.0.0
fastapi==0.105.0
favicon==0.7.0
fonttools==4.47.0
gitdb==4.0.11
GitPython==3.1.40
h11==0.14.0
htbuilder==0.6.2
httptools==0.6.1
idna==3.6
importlib-metadata==6.11.0
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
kiwisolver==1.4.5
lxml==4.9.4
Markdown==3.5.1
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.3
matplotlib==3.8.2
mdurl==0.1.2
more-itertools==10.1.0
numpy==1.25
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
pipreqs==0.4.13
prettytable==3.9.0
protobuf==4.25.1
pyarrow==14.0.2
pydantic==2.5.2
pydantic_core==2.14.5
pydeck==0.8.1b0
pyecharts==2.0.4
Pygments==2.17.2
pymdown-extensions==10.5
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
requests==2.31.0
rich==13.7.0
rpds-py==0.15.2
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
st-annotated-text==4.0.1
starlette==0.27.0
streamlit==1.29.0
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.0
streamlit-echarts==0.4.0
streamlit-embedcode==0.1.2
streamlit-extras==0.3.6
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.6
streamlit-keyup==0.2.2
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.0.5
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.4
typing_extensions==4.9.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.1.0
uvicorn==0.24.0.post1
validators==0.22.0
watchdog==3.0.0
watchfiles==0.21.0
wcwidth==0.2.12
websockets==12.0
yarg==0.1.9
zipp==3.17.0
plotly==5.18.0
ydata_profiling==4.6.3
streamlit_pandas_profiling==0.1.3